Takistuse massiiv on 4-kohaline ehk vahed on iga posti vahel samad, kommentaarid, kiiruse ja gravitatsiooni korrigeerimine.

+++ src/Mandariin.java
@@ -1,16 +1,16 @@
 import java.awt.*;
 
 public class Mandariin {
-    private int mängijaAsukoht;
-    private int gravitatsioon;
-    private int kiirus;
+    private int mängijaAsukoht; // Mängija keskpunkti koordinaadid
+    private double gravitatsioon; // Kukkumise kiirendus
+    private double kiirus; // Hetkeline kiirus, mida muudetakse hüppamise ja kukkumisega
     private int mängijaDiameeter;
 
     public int getMängijaAsukoht() {
         return mängijaAsukoht;
     }
 
-    public int getGravitatsioon() {
+    public double getGravitatsioon() {
         return gravitatsioon;
     }
 
@@ -18,15 +18,15 @@ public class Mandariin {
         return mängijaDiameeter;
     }
 
-    public int getKiirus() {
+    public double getKiirus() {
         return kiirus;
     }
 
-    public void setKiirus(int kiirus) {
+    public void setKiirus(double kiirus) {
         this.kiirus = kiirus;
     }
 
-    public void setGravitatsioon(int gravitatsioon) {
+    public void setGravitatsioon(double gravitatsioon) {
         this.gravitatsioon = gravitatsioon;
     }
 
@@ -34,14 +34,14 @@ public class Mandariin {
         this.mängijaAsukoht = mängijaAsukoht;
     }
 
-    public Mandariin(int mängijaAsukoht, int gravitatsioon, int kiirus, int mängijaDiameeter) {
+    public Mandariin(int mängijaAsukoht, double gravitatsioon, double kiirus, int mängijaDiameeter) {
         this.mängijaAsukoht = mängijaAsukoht;
         this.gravitatsioon = gravitatsioon;
         this.mängijaDiameeter = mängijaDiameeter;
         this.kiirus = kiirus;
     }
 
-    public void muudaKiirust(int uusKiirus){
+    public void muudaKiirust(double uusKiirus){
         this.kiirus = uusKiirus;
     }
 }
